examples/bitcoin-tracker/requirements.txt
requests>=2.32.5
google-cloud-bigquery>=3.11.0
